Code Cleanup
Updating to VS2019.
Changing to .NET  Core Preview.
Removed broken .NET Framework Console project.
Started adding Default Signature methods which should point to the best option in the test class for reference from other methods.
Cleaned and fixed up the Primitive Structs.
Added 4D versions of Primitive Structs.
Cleaned and fixed up Shape Primitive Structs.
Filled in missing faster Bezier Coefficient methods up through Decic.
Moved more Code from Engine Test Project.

+++ InstrumentedLibrary/Framework/Enums/Signs.cs
@@ -4,7 +4,7 @@
     /// The signs enum.
     /// </summary>
     public enum Signs
-        : int
+        : sbyte
     {
         /// <summary>
         /// The Number is Negative = -1.
